requirements.txt
# Self-Learning AI Agent requirements
typer>=0.9.0
rich>=13.0.0
watchdog>=3.0.0
pyyaml>=6.0
gitpython>=3.1.40
pathspec>=0.11.0
